+++ src/main/java/Ingredient.java
@@ -6,7 +6,7 @@ public class Ingredient {
     private Integer id;
     private String name;
     private CategoryEnum category;
-    private BigDecimal price;
+    private Double price;
 
     @Override
     public boolean equals(Object o) {
@@ -42,7 +42,7 @@ public class Ingredient {
         return category;
     }
 
-    public BigDecimal getPrice() {
+    public Double getPrice() {
         return price;
     }
 
@@ -58,11 +58,11 @@ public class Ingredient {
         this.category = category;
     }
 
-    public void setPrice(BigDecimal price) {
+    public void setPrice(Double price) {
         this.price = price;
     }
 
-    public Ingredient(Integer id, String name, CategoryEnum category, BigDecimal price) {
+    public Ingredient(Integer id, String name, CategoryEnum category, Double price) {
         this.id = id;
         this.name = name;
         this.category = category;
